requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via kopf
aiosignal==1.2.0
    # via aiohttp
asttokens==2.0.5
    # via stack-data
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
backcall==0.2.0
    # via ipython
black==22.3.0
    # via -r requirements.in
cachetools==5.1.0
    # via google-auth
certifi==2022.5.18.1
    # via
    #   kubernetes
    #   requests
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   black
    #   kopf
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
executing==0.8.3
    # via stack-data
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.6.6
    # via kubernetes
idna==3.3
    # via
    #   requests
    #   yarl
ipdb==0.13.9
    # via -r requirements.in
ipython==8.3.0
    # via
    #   -r requirements.in
    #   ipdb
iso8601==1.0.2
    # via kopf
jedi==0.18.1
    # via ipython
kopf==1.35.4
    # via -r requirements.in
kubernetes==23.6.0
    # via -r requirements.in
matplotlib-inline==0.1.3
    # via ipython
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via black
oauthlib==3.2.0
    # via requests-oauthlib
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.5.2
    # via black
prompt-toolkit==3.0.29
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pygments==2.12.0
    # via ipython
python-dateutil==2.8.2
    # via kubernetes
python-json-logger==2.0.2
    # via kopf
pyyaml==6.0
    # via
    #   -r requirements.in
    #   kopf
    #   kubernetes
requests==2.27.1
    # via
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   asttokens
    #   google-auth
    #   kubernetes
    #   python-dateutil
stack-data==0.2.0
    # via ipython
toml==0.10.2
    # via ipdb
tomli==2.0.1
    # via black
traitlets==5.2.1.post0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.2.0
    # via kopf
urllib3==1.26.9
    # via
    #   kubernetes
    #   requests
wcwidth==0.2.5
    # via prompt-toolkit
websocket-client==1.3.2
    # via kubernetes
yarl==1.7.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
